+++ docs/src/introduction/get-started.js
@@ -1,4 +1,5 @@
 import React from "react";
+import MarkdownShard from "@fa-repo/shard-docs/dist/shards/markdown-shard";
 import SectionShard from "@fa-repo/shard-docs/dist/shards/section-shard";
 import CodeExampleShard from "@fa-repo/shard-docs/dist/shards/code-example-shard";
 
@@ -8,29 +9,37 @@ import CodeExampleShard from "@fa-repo/shard-docs/dist/shards/code-example-shard
 
 const GetStartedPage = [
   <SectionShard title="Installation" persistState="35759">
-    <CodeExampleShard lang="bash" noShadow sourceCode="npm install @fa-repo/shard-docs" />
+    <MarkdownShard
+      lang="bash"
+      markdown={`\`\`\`jsx
+npm install @fa-repo/shard-docs
+\`\`\`
+`}
+    />
   </SectionShard>,
   <SectionShard title="Dependencies" persistState="57226">
     <p>
       ShardDocs makes use of React Router V4. Ensure that an instance of this component is wrapped
       by a router provider further up the tree. Example:
     </p>
-    <CodeExampleShard
-      noShadow
-      sourceCode={`import { BrowserRouter } from "react-router-dom";
+    <MarkdownShard
+      markdown={`\`\`\`jsx
+import { BrowserRouter } from "react-router-dom";
 
 render(
   <BrowserRouter>
     <ShardDocs title="Documentation title" />
   </BrowserRouter>,
   document.getElementById("app")
-);`}
+);
+\`\`\`
+`}
     />
   </SectionShard>,
   <SectionShard title="Basic usage" persistState="33306">
-    <CodeExampleShard
-      noShadow
-      sourceCode={`ReactDOM.render(
+    <MarkdownShard
+      markdown={`\`\`\`jsx
+ReactDOM.render(
   <ShardDocs
     title="Documentation title"
     structure={[
@@ -39,7 +48,9 @@ render(
     ]}
   />,
     document.getElementById('app')
-);`}
+);
+\`\`\`
+`}
     />
   </SectionShard>
 ];